New dataset generator and associated training script for cityscapes.
In this case we use directly files to speed up the training.
When using h5py, we had some problems due to multiprocessing of keras.
When  using the on-the-fly calculation of mask based on the polygons it took too much time.
These are the reasons to save the dataset as files. There 2 types of files added to the annotation directory of each city.
One file is the instance image, which each pixel contains a color that represents the associated instance (each instance has different color), and a label csv file that
contains in each line the r,g,b and label for each instance in this instance image.